+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added `knitr` engine `rtex` (see vignette "Chunk engine for easy tex output in Rmarkdown and Quarto")
 * New functions: `as_expr.caracas_solve_sys_sol`
 * `inv()` by default uses `gauss` like SymPy
+* `kronecker`, `rep` added
 
 # caracas 2.0.1